fix(cond_transition): bf16 correctness for fused inference kernel (v1)
The cond_transition triton kernels were fp32/TF32-only. `_cond_transition_inference_kernel`
builds `h = a*sigmoid(a)*b` in fp32 then `tl.dot(h, ws_t)`; in bf16 `ws_t` is bf16 while `h`
is fp32 -> CompilationError (eager) or a NaN graph-break fallback (under torch.compile). The
module bench hid this: squeeze uses init="zero" (output identically 0) and the bench computes
no output_cosine, so a NaN/broken kernel "passed".

Hypothesis: casting the SwiGLU intermediate to the operand dtype before the squeeze dot makes
it a proper bf16 MMA (no-op for fp32).

Fix: `h = (a * tl.sigmoid(a) * b).to(x.dtype)`.

Measured (B200, A48, non-zero squeeze, eager): bf16 inference cos 0.999987 (was NaN);
fp32 inference cos 1.0 unchanged. Deployment regime (compile+cudagraph) speedup vs compiled
torch: 1.02x @L1024, 1.09x @L4096 (timing unchanged by the cast; the win is now real).

Validation: python ncu/ct_matrix.py

+++ src/miniworld_kernels/kernels/conditioned_transition/triton/inference.py
@@ -76,7 +76,7 @@ def _cond_transition_inference_kernel(
         )
         a = tl.dot(x, wa, out_dtype=tl.float32, input_precision="tf32")
         b = tl.dot(x, wb, out_dtype=tl.float32, input_precision="tf32")
-        h = a * tl.sigmoid(a) * b  # (BM, BN) fp32
+        h = (a * tl.sigmoid(a) * b).to(x.dtype)  # cast to operand dtype -> squeeze dot works in bf16 (no-op for fp32)
         ws_t = tl.load(  # (BN, D): Ws[d, cols]^T
             ws_ptr + cols[:, None] * stride_sn + dcols[None, :] * stride_sd,
             mask=col_mask[:, None], other=0.0,
